Fix chat room state management
