+++ src/api.js
@@ -2,6 +2,7 @@ import LivechatClient from '@rocket.chat/sdk/lib/clients/Livechat';
 
 export let Livechat;
 
-export const initializeLivechat = (...args) => {
+export const initializeLivechat = async(...args) => {
 	Livechat = new LivechatClient(...args);
+	Livechat.connect();
 };
